bidirectional compilation
rewrote the entire xml -> nl module to use the up to date backwards comilation structure.
this gives us the ability to compile back and forth between netsblox netlogo.
there are some minor issues that need to be resolved (like redundant let and set from var decl), but they should be easy to fix.