added hatching
added hatch statements.
refactored the old "%n clones of %s" to "%n clones" to do this.
this is great, cause before the clones of block was a framework tool that users weren't allowed to use.

fixed an issue where we were potentially mutating sprite sets while iterating in a tell block.

added support for get/set heading.

added black and white to colors, which i missed before.